fix: phpstan error
